arr3

+++ 23_arrays/ArrayTest1.java
@@ -1,6 +1,6 @@
 public class ArrayTest1 {
     public static void main(String[] args) {
-        short [] lottonumbers = new short[6];
+        float [] lottonumbers = new float[6];
        
         System.out.println("value = "+lottonumbers[0]);
